+++ components/DeleteButton.js
@@ -1,0 +1,13 @@
+import {IconButton} from 'react-native-paper';
+
+export default function DeleteButton({onPress}) {
+  return (
+    <IconButton
+      iconColor="white"
+      icon="delete"
+      mode="default"
+      size={30}
+      onPress={onPress}
+    />
+  );
+}
