+++ components/ItemCard.js
@@ -1,7 +1,9 @@
 import * as React from 'react';
 import {StyleSheet, TouchableOpacity, View} from 'react-native';
 import LinearGradient from 'react-native-linear-gradient';
-import {Card, Text, IconButton} from 'react-native-paper';
+import {Card, Text} from 'react-native-paper';
+import UpdateButton from './UpdateButton';
+import DeleteButton from './DeleteButton';
 
 function ItemCard({id, name, icon, handleDelete, onPress}) {
   const airlineName = name.length > 15 ? name.substring(0, 16) + '...' : name;
@@ -28,11 +30,22 @@ function ItemCard({id, name, icon, handleDelete, onPress}) {
                 </View>
               </Card.Content>
               <Card.Actions>
+                <UpdateButton onPress={() => handleUpdate(id, name)} />
+                <DeleteButton onPress={() => handleDelete(id, name)} />
+                {/* <IconButton
+                  icon="pencil"
+                  mode="default"
+                  iconColor="white"
+                  size={30}
+                  onPress={() => handleUpdate(id, name)}
+                />
                 <IconButton
+                  iconColor="white"
                   icon="delete"
+                  mode="default"
                   size={30}
                   onPress={() => handleDelete(id, name)}
-                />
+                /> */}
               </Card.Actions>
             </View>
           </LinearGradient>
